backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
google-generativeai==0.7.2
Pillow>=10.0.0
pytesseract==0.3.10
